fix: eventstream invalid topic error
