+++ packages/api/src/utils/server/index.ts
@@ -1,3 +1,4 @@
 export * from "./genericJsonServer.js";
 export * from "./registerRoute.js";
+export * from "./errors.js";
 export * from "./types.js";
